Big bunch of cleanups. 1. Refactored a bunch of the LstObjectFileLoader code. 2. Converted the SOURCE stuff into a couple of classes instead of the map for each PObject.  Should save a few MB. 3. All the usual 1.5 conversions and JavaDoc.
git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@1251 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/pcgen/gui/MainHP.java
@@ -474,7 +474,7 @@ final class MainHP extends JPanel
 					{
 						final int lvl = aPC.getLevelInfoClassLevel(rowIndex);
 						final int baseSides = aClass.getLevelHitDie(aPC, lvl);
-						iHp = aClass.getHitPoint(lvl - 1).intValue();
+						iHp = aClass.getHitPoint(lvl - 1);
 						iSides = baseSides + (int) aClass.getBonusTo("HD", "MAX", lvl, aPC);
 
 						//
